Add useEffect to randomize cards and duplicate card function
Added the shuffleCards function to useEffect so the cards are randomized
on every render and mount. Created the checkDuplicate function to check
if a card has been selected twice. Resets score and sets highscore if
appropriate if true. Adds one to score if not. Added several images to
be used instead of placeholder letters.